Move initialValues and schema inside Form
