+++ src/screens/authentication/SignUp/index.js
@@ -1,23 +1,8 @@
 import React, { useState } from 'react'
-import * as Yup from 'yup'
 import NavigationService from '~/navigationService'
 import * as AnalyticsService from '~/analyticsService'
 import * as AuthenticationService from '~/authenticationService'
 import Auth from '../components/Auth'
-
-const initialValues = {
-  credential: '',
-  password: ''
-}
-
-const schema = Yup.object().shape({
-  credential: Yup.string()
-    .email('Email inválido')
-    .required('Requerido'),
-  password: Yup.string()
-    .min(6, 'Contraseña muy corta')
-    .required('Requerido')
-})
 
 const SignUp = () => {
   const [authenticating, setAuthenticating] = useState(false)
@@ -42,13 +27,12 @@ const SignUp = () => {
       <Auth.Navigation goTo='Login'>Iniciar sesión</Auth.Navigation>
       <Auth.Title>Creá tu cuenta</Auth.Title>
       <Auth.Form
-        initialValues={initialValues}
         handleSubmit={handleSubmit}
-        schema={schema}
         error={error}
         authenticating={authenticating}
         credentialText='Ingresá tu email'
         submitText='Registrarme'
+        applyValidation
       />
       <Auth.Separator>Creá tu cuenta usando</Auth.Separator>
       <Auth.SocialNetworks />
